problem: import account from Ledger Nano crashes solution: When select account from store use only lowcase id
