+++ src/components/ledger/ImportAccount/addrlist.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Table, TableBody, TableHeader, TableHeaderColumn, TableRow } from 'material-ui/Table';
 import { tables } from 'lib/styles';
-import ledger from 'store/ledger';
+import ledger from '../../../store/ledger';
 import Addr from './addr';
 
 class AddrList extends React.Component {
